move fragments to an own package
