+++ app/src/main/java/es/wolfi/app/passman/activities/PasswordListActivity.java
@@ -61,16 +61,16 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.Objects;
 
-import es.wolfi.app.passman.CredentialAdd;
-import es.wolfi.app.passman.CredentialDisplay;
-import es.wolfi.app.passman.CredentialEdit;
-import es.wolfi.app.passman.CredentialItemFragment;
+import es.wolfi.app.passman.fragments.CredentialAdd;
+import es.wolfi.app.passman.fragments.CredentialDisplay;
+import es.wolfi.app.passman.fragments.CredentialEdit;
+import es.wolfi.app.passman.fragments.CredentialItemFragment;
 import es.wolfi.app.passman.R;
 import es.wolfi.app.passman.SettingValues;
-import es.wolfi.app.passman.Settings;
+import es.wolfi.app.passman.fragments.Settings;
 import es.wolfi.app.passman.SingleTon;
-import es.wolfi.app.passman.VaultFragment;
-import es.wolfi.app.passman.VaultLockScreen;
+import es.wolfi.app.passman.fragments.VaultFragment;
+import es.wolfi.app.passman.fragments.VaultLockScreen;
 import es.wolfi.passman.API.Core;
 import es.wolfi.passman.API.Credential;
 import es.wolfi.passman.API.File;
