+++ app/src/main/java/es/wolfi/app/passman/fragments/VaultLockScreen.java
@@ -20,7 +20,7 @@
  *
  */
 
-package es.wolfi.app.passman;
+package es.wolfi.app.passman.fragments;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -41,6 +41,9 @@ import android.widget.TextView;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
+import es.wolfi.app.passman.R;
+import es.wolfi.app.passman.SettingValues;
+import es.wolfi.app.passman.SingleTon;
 import es.wolfi.passman.API.Vault;
