+++ app/src/main/java/es/wolfi/utils/FilterListAsyncTask.java
@@ -28,9 +28,9 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
 
-import es.wolfi.app.passman.CredentialItemFragment;
+import es.wolfi.app.passman.fragments.CredentialItemFragment;
 import es.wolfi.app.passman.adapters.CredentialViewAdapter;
-import es.wolfi.app.passman.VaultFragment;
+import es.wolfi.app.passman.fragments.VaultFragment;
 import es.wolfi.app.passman.adapters.VaultViewAdapter;
 import es.wolfi.passman.API.Credential;
 import es.wolfi.passman.API.Vault;
