+++ app/src/main/res/layout/fragment_credential_add.xml
@@ -37,7 +37,7 @@
             android:paddingTop="@dimen/activity_vertical_margin"
             android:paddingRight="@dimen/activity_horizontal_margin"
             android:paddingBottom="@dimen/floating_button_parent_padding"
-            tools:context="es.wolfi.app.passman.CredentialDisplay">
+            tools:context="es.wolfi.app.passman.fragments.CredentialDisplay">
 
             <TextView
                 android:id="@+id/add_credential_label_header"
@@ -133,7 +133,7 @@
                 android:layout_weight="1"
                 android:nestedScrollingEnabled="false"
                 app:layoutManager="LinearLayoutManager"
-                tools:context="es.wolfi.app.passman.CredentialAdd"
+                tools:context="es.wolfi.app.passman.fragments.CredentialAdd"
                 tools:itemCount="2"
                 tools:listitem="@layout/fragment_credential_file_item">
 
@@ -161,7 +161,7 @@
                 android:layout_weight="1"
                 android:nestedScrollingEnabled="false"
                 app:layoutManager="LinearLayoutManager"
-                tools:context="es.wolfi.app.passman.CredentialEdit"
+                tools:context="es.wolfi.app.passman.fragments.CredentialEdit"
                 tools:itemCount="2"
                 tools:listitem="@layout/fragment_credential_custom_field_item"></androidx.recyclerview.widget.RecyclerView>
