+++ app/src/main/res/layout/fragment_credential_display.xml
@@ -33,7 +33,7 @@
         android:paddingLeft="@dimen/activity_horizontal_margin"
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
-        tools:context="es.wolfi.app.passman.CredentialDisplay">
+        tools:context="es.wolfi.app.passman.fragments.CredentialDisplay">
 
         <TextView
             style="@style/Label"
@@ -147,7 +147,7 @@
             android:nestedScrollingEnabled="false"
             app:layoutManager="LinearLayoutManager"
             tools:itemCount="2"
-            tools:context="es.wolfi.app.passman.CredentialDisplay"
+            tools:context="es.wolfi.app.passman.fragments.CredentialDisplay"
             tools:listitem="@layout/fragment_credential_file_item">
         </androidx.recyclerview.widget.RecyclerView>
 
@@ -165,7 +165,7 @@
             android:nestedScrollingEnabled="false"
             app:layoutManager="LinearLayoutManager"
             tools:itemCount="2"
-            tools:context="es.wolfi.app.passman.CredentialDisplay"
+            tools:context="es.wolfi.app.passman.fragments.CredentialDisplay"
             tools:listitem="@layout/fragment_credential_custom_field_item">
         </androidx.recyclerview.widget.RecyclerView>
