+++ app/src/main/res/layout/fragment_credential_item_list.xml
@@ -57,7 +57,7 @@
         android:layout_height="match_parent"
         android:layout_weight="1"
         app:layoutManager="LinearLayoutManager"
-        tools:context="es.wolfi.app.passman.CredentialItemFragment"
+        tools:context="es.wolfi.app.passman.fragments.CredentialItemFragment"
         tools:listitem="@layout/fragment_credential_item">
 
     </androidx.recyclerview.widget.RecyclerView>
