+++ app/src/main/res/layout/fragment_settings.xml
@@ -38,7 +38,7 @@
             android:paddingTop="@dimen/activity_vertical_margin"
             android:paddingRight="@dimen/activity_horizontal_margin"
             android:paddingBottom="@dimen/activity_vertical_margin"
-            tools:context="es.wolfi.app.passman.CredentialDisplay">
+            tools:context="es.wolfi.app.passman.fragments.CredentialDisplay">
 
             <TextView
                 style="@style/TextAppearance.AppCompat.Medium"
