+++ app/src/main/res/layout/fragment_vault_list.xml
@@ -30,5 +30,5 @@
     android:layout_marginLeft="0dp"
     android:layout_marginRight="0dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context="es.wolfi.app.passman.VaultFragment"
+    tools:context="es.wolfi.app.passman.fragments.VaultFragment"
     tools:listitem="@layout/fragment_vault" />
