+++ app/src/main/res/layout/fragment_vault_lock_screen.xml
@@ -27,7 +27,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="es.wolfi.app.passman.VaultLockScreen">
+    tools:context="es.wolfi.app.passman.fragments.VaultLockScreen">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
